--- a/benchmark.py
+++ b/benchmark.py
@@ -53,20 +53,20 @@ if __name__ == "__main__":
     # Profiling info about ops are saved in 'test-%s.txt' % FLAGS.out
     profiler.profile_operations(options=opts)
 
-    # for batch_size in [2 ** n for n in range(8)]:
-    #     # Init data
-    #     points_with_colors = np.random.randn(batch_size, hyper_params["num_point"], 6)
-    #
-    #     # Warm up
-    #     pd_labels = predictor.predict(points_with_colors)
-    #
-    #     # Benchmark
-    #     s = time.time()
-    #     _ = predictor.predict(points_with_colors)
-    #     batch_time = time.time() - s
-    #     sample_time = batch_time / batch_size
-    #     print(
-    #         "Batch size: {}, batch_time: {}, sample_time: {}".format(
-    #             batch_size, batch_time, sample_time
-    #         )
-    #     )
+    for batch_size in [2 ** n for n in range(8)]:
+        # Init data
+        points_with_colors = np.random.randn(batch_size, hyper_params["num_point"], 6)
+
+        # Warm up
+        pd_labels = predictor.predict(points_with_colors)
+
+        # Benchmark
+        s = time.time()
+        _ = predictor.predict(points_with_colors)
+        batch_time = time.time() - s
+        sample_time = batch_time / batch_size
+        print(
+            "Batch size: {}, batch_time: {}, sample_time: {}".format(
+                batch_size, batch_time, sample_time
+            )
+        )