--- a/tf_ops/test_interpolate.py
+++ b/tf_ops/test_interpolate.py
@@ -1,47 +1,35 @@
 import numpy as np
 import tensorflow as tf
+import time
 from tf_interpolate import three_nn, three_interpolate
 
 if __name__ == "__main__":
     np.random.seed(100)
-    pts = np.random.random((32, 128, 64)).astype("float32")
-    tmp1 = np.random.random((32, 512, 3)).astype("float32")
-    tmp2 = np.random.random((32, 128, 3)).astype("float32")
+
+    target_points = np.random.random((64, 8192, 3)).astype("float32")
+    base_points = np.random.random((64, 1024, 3)).astype("float32")
 
     with tf.device("/cpu:0"):
-        points = tf.constant(pts)
-        xyz1 = tf.constant(tmp1)
-        xyz2 = tf.constant(tmp2)
+        xyz1 = tf.constant(target_points)
+        xyz2 = tf.constant(base_points)
         dist, idx = three_nn(xyz1, xyz2)
 
     with tf.Session("") as sess:
+        # Warm up
         dist, idx = sess.run(three_nn(xyz1, xyz2))
-        print(dist.shape, dist.dtype)
-        print(dist[:3, :3, :1])
-        print(idx.shape, idx.dtype)
-        print(idx[:3, :3, :1])
 
-    # (32, 512, 3) float32
-    # [[[0.03415794]
-    # [0.01955329]
-    # [0.00292144]]
-    #
-    # [[0.0038327 ]
-    # [0.01290642]
-    # [0.01430019]]
-    #
-    # [[0.04590415]
-    # [0.01247532]
-    # [0.00384166]]]
-    # (32, 512, 3) int32
-    # [[[ 62]
-    # [ 83]
-    # [ 53]]
-    #
-    # [[ 71]
-    # [ 82]
-    # [112]]
-    #
-    # [[ 77]
-    # [126]
-    # [ 42]]]
+        # Run
+        s = time.time()
+        dist, idx = sess.run(three_nn(xyz1, xyz2))
+        print("Time: {}".format(time.time() - s))
+        print(dist[:3, :3, :1].flatten())
+        print(idx[:3, :3, :1].flatten())
+        print(idx.shape, idx.dtype)
+        print(dist.shape, dist.dtype)
+
+        # Expected output
+        # [0.00175864 0.00671887 0.0034472  0.00337327 0.00191902 0.00075543
+        #  0.00169418 0.00473733 0.00381071]
+        # [137 856 116  76 915 199 117 659 786]
+        # (64, 8192, 3) int32
+        # (64, 8192, 3) float32