--- a/tf_ops/tf_interpolate.cpp
+++ b/tf_ops/tf_interpolate.cpp
@@ -53,9 +53,18 @@ static double get_time() {
     return tp.tv_sec + tp.tv_nsec * 1e-9;
 }
 
-// Find three nearest neigbors with square distance
+// Find three nearest neighbors with square distance
 // input: xyz1 (b,n,3), xyz2(b,m,3)
 // output: dist (b,n,3), idx (b,n,3)
+// E.g.
+// - target_points (b, n, 3): (64, 8192, 3), the "3" here is x, y, z
+// - base_points (b, m, 3): (64, 256, 3), the "3" here is x, y, z
+// - dist (b, n, 3): (64, 8192, 3), for each input point in target_points, find
+//                   3 nearest neighbors in base_points and return the distances,
+//                   the "3" means "3" nearest neighbors
+// - idx (b, n, 3): (64, 8192, 3), for each input point in target_points, find 3
+//                   nearest neighbors in base_points and return the indexes in
+//                   base_points, the "3" means "3" nearest neighbors
 void threenn_cpu(int b, int n, int m, const float *xyz1, const float *xyz2,
                  float *dist, int *idx) {
     for (int i = 0; i < b; ++i) {