--- a/tf_ops/tf_interpolate.cpp
+++ b/tf_ops/tf_interpolate.cpp
@@ -58,8 +58,8 @@ static double get_time() {
 // input: xyz1 (b,n,3), xyz2(b,m,3)
 // output: dist (b,n,3), idx (b,n,3)
 // E.g.
-// - target_points (b, n, 3): (64, 8192, 3), the "3" here is x, y, z
-// - base_points (b, m, 3): (64, 256, 3), the "3" here is x, y, z
+// - target_points (b, n, 3): e.g. (64, 8192, 3), the "3" here is x, y, z
+// - base_points (b, m, 3): e.g. (64, 1024, 3), the "3" here is x, y, z
 // - dist (b, n, 3): (64, 8192, 3), for each input point in target_points, find
 //                   3 nearest neighbors in base_points and return the
 //                   distances, the "3" means "3" nearest neighbors