About 5x speed up of the `ThreeNN` custom TF op, about 3x speed up for the full prediction model.

Benchmark with `benchmark.py`

## Before
Running full model prediction
```
Batch size: 1, batch_time: 0.11297130584716797, sample_time: 0.11297130584716797
Batch size: 2, batch_time: 0.18468594551086426, sample_time: 0.09234297275543213
Batch size: 4, batch_time: 0.3397715091705322, sample_time: 0.08494287729263306
Batch size: 8, batch_time: 0.6424834728240967, sample_time: 0.08031043410301208
Batch size: 16, batch_time: 1.0457987785339355, sample_time: 0.06536242365837097
Batch size: 32, batch_time: 2.30259370803833, sample_time: 0.07195605337619781
Batch size: 64, batch_time: 4.3022308349609375, sample_time: 0.06722235679626465
Batch size: 128, batch_time: 8.371811151504517, sample_time: 0.0654047746211290
```

Running benchmark by op, batch_size = 64
```
node name | requested bytes | total execution time | accelerator execution time | cpu execution time | op occurrence (run|defined)
ThreeNN | 14.65MB (100.00%, 0.11%), | 4.21sec (100.00%, 87.94%), | 0us (100.00%, 0.00%), | 4.21sec (100.00%, 91.42%), | 4|4
```

## After
Running full model prediction
```
Batch size: 1, batch_time: 0.044135332107543945, sample_time: 0.044135332107543945
Batch size: 2, batch_time: 0.06282544136047363, sample_time: 0.031412720680236816
Batch size: 4, batch_time: 0.10235333442687988, sample_time: 0.02558833360671997
Batch size: 8, batch_time: 0.18165111541748047, sample_time: 0.02270638942718506
Batch size: 16, batch_time: 0.328458309173584, sample_time: 0.020528644323349
Batch size: 32, batch_time: 0.5884418487548828, sample_time: 0.018388807773590088
Batch size: 64, batch_time: 1.1684465408325195, sample_time: 0.018256977200508118
Batch size: 128, batch_time: 2.2928593158721924, sample_time: 0.017912963405251503
```

Running benchmark by op, batch_size = 64
```
node name | requested bytes | total execution time | accelerator execution time | cpu execution time | op occurrence (run|defined)
ThreeNN | 14.65MB (100.00%, 0.11%), | 848.40ms (100.00%, 59.31%), | 0us (100.00%, 0.00%), | 848.40ms (100.00%, 67.42%), | 4|4
```

## If OpenMP enabled
```
Batch size: 1, batch_time: 0.13734221458435059, sample_time: 0.13734221458435059
Batch size: 2, batch_time: 0.17317986488342285, sample_time: 0.08658993244171143
Batch size: 4, batch_time: 0.09513401985168457, sample_time: 0.023783504962921143
Batch size: 8, batch_time: 0.2525460720062256, sample_time: 0.0315682590007782
Batch size: 16, batch_time: 0.3754005432128906, sample_time: 0.023462533950805664
Batch size: 32, batch_time: 0.5127174854278564, sample_time: 0.016022421419620514
Batch size: 64, batch_time: 0.781425952911377, sample_time: 0.012209780514240265
Batch size: 128, batch_time: 1.4844071865081787, sample_time: 0.011596931144595146
```