@syncle 2018-12-14
nice! about 5x faster